cleaned up unnecessary files, renamed and moved stuff, and linted
