+++ styles/NewsletterCarouselCard.module.css
@@ -1,49 +1,3 @@
-/* .content {
-  display: grid;
-  gap: 2rem;
-  justify-items: center;
-  width: 100%;
-  grid-template-columns: 500px 1fr;
-  width: 1024px;
-  padding: 20px;
-}
-
-.news-poster {
-  position: relative;
-  width: 100%;
-  margin-bottom: 1.5rem;
-  border-radius: 10px;
-}
-
-.news-poster > img {
-  right: 0 !important;
-  left: auto !important;
-  max-width: 100%;
-  max-height: 100%;
-  width: auto !important;
-}
-
-.content-column {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  padding-bottom: 1rem;
-  margin-bottom: 1.5rem;
-  gap: 10px;
-}
-
-
-
-
-.text-content {
-  display: flex;
-  flex-direction: column;
-  gap: 2.5rem;
-  text-align: left;
-  color: white;
-}
- */
-
 .container {
   display: flex;
   justify-content: center;
@@ -112,7 +66,7 @@
     float: left;
     display: block;
     padding: 0 10px;
-}
+  }
   .container {
     height: auto;
   }
